[WORKAROUND] Change jv.offset field to 32 bit size
Related issues:
  - Issue #1108  surprising bug if n > 65535
    https://github.com/stedolan/jq/issues/1108

+++ src/jv.h
@@ -23,7 +23,7 @@ struct jv_refcnt;
 typedef struct {
   unsigned char kind_flags;
   unsigned char pad_;
-  unsigned short offset;  /* array offsets */
+  unsigned int offset;  /* array offsets */
   int size;
   union {
     struct jv_refcnt* ptr;
